DF-24.1: Knowledge physical cleanup — rosclaw.know → rosclaw.knowledge.legacy
Phase-II DF-16.1 (first of the three-way physical-move split):

- git mv know → knowledge/legacy; DF-09 alias module dissolved into the
  package __init__ (LegacyKnowledgeRuntime + cards merged with existing
  exports)
- intra-package imports rewritten to the canonical path; logger names
  kept for log-grep compat
- rosclaw.know stays as deprecation shim (package re-export + 12
  submodule shims aliasing sys.modules — identity, not copies)
- all in-repo importers migrated; source-discipline test pins zero shim
  imports inside src
- tests/knowledge/test_physical_move_shim.py (4); 391 affected +
  full regression 6924 passed / 29 host-baseline flakes (same as main)
