+++ src/rosclaw/knowledge/legacy/assets/failure_taxonomy.yaml
