+++ src/rosclaw/simforge/contact_push_learning.py
@@ -11,7 +11,7 @@ from typing import Any
 
 import numpy as np
 
-from rosclaw.know.task_card import TaskCard
+from rosclaw.knowledge.legacy.task_card import TaskCard
 from rosclaw.memory.seekdb_client import InMemoryStructuredStore
 from rosclaw.memory.v2.models import MemoryItem, MemoryStatus, MemoryType
 from rosclaw.memory.v2.repository import MemoryRepository
