+++ tests/test_practice_v2_evidence_bridge.py
@@ -8,7 +8,7 @@ from types import SimpleNamespace
 
 from rosclaw.cli import cmd_practice_record
 from rosclaw.how.engine import HeuristicEngine
-from rosclaw.know.interface import KnowledgeInterface
+from rosclaw.knowledge.legacy.interface import KnowledgeInterface
 from rosclaw.memory.interface import MemoryInterface
 from rosclaw.memory.seekdb_client import InMemoryKnowledgeStore
